+++ iow/workflow/workflow_options.hpp
@@ -14,7 +14,7 @@ struct workflow_options
   // в delayed_queue есть баг
   bool use_io_service = true;
   bool debug = false;
-  int  threads = 0;
+  size_t  threads = 0;
 
   size_t maxsize = 1024*1024*128;
   size_t wrnsize = 1024*1024;
